V1.1.1 - Minor bug changes
Output showing as expected (decimal 2 digits)

+++ Desafio_Nubank/Models/Operation.cs
@@ -6,7 +6,7 @@ class Operation
     public OperationType Type { get; set; }
 
     [JsonPropertyName("unit-cost")]
-    public double UnitCost { get; set; }
+    public decimal UnitCost { get; set; }
 
     [JsonPropertyName("quantity")]
     public int Quantity { get; set; }
